requirements.txt
selenium
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
pymysql
openpyxl
jsonpath
requests
rsa
